Change: more resolution to timestamp
Use microsecond to timestamp to make it more precise. Make
corresponding changes in Java class.

+++ src/org/isoblue/can/CanSocketJ1939.java
@@ -116,7 +116,7 @@ public class CanSocketJ1939 extends CanSocket {
 		protected int len;
 		protected int priority;
 		protected byte[] data;
-		protected int timestamp;
+		protected float timestamp;
     		
 		private static String byteArrayToHex(byte[] a) {
        			StringBuilder sb = new StringBuilder(a.length * 2);
@@ -127,7 +127,7 @@ public class CanSocketJ1939 extends CanSocket {
 
 		public void print(final int verbose) {
 			if (verbose == 1) {
-				System.out.printf("\n%d,%s,%d,%d,%d,%d,%d,%d,"
+				System.out.printf("\n%.3f,%s,%d,%d,%d,%d,%d,%d,"
 					+ "%d,%s", timestamp, ifName, name,
 					addr, dstName, dstAddr, pgn, len,
 					priority, byteArrayToHex(data));
@@ -142,7 +142,7 @@ public class CanSocketJ1939 extends CanSocket {
 		/* recv frame constructor */
 		public Message(String ifName, long name, int addr,
 			long dstName, int dstAddr, int pgn, int len,
-			int priority, byte[] data, int timestamp) {
+			int priority, byte[] data, float timestamp) {
 			this.ifName = ifName;
 			this.name = name;
 			this.addr = addr;
